Fixed sync_list creation

+++ lliurex-onedrive.install/gui/OnedriveManager.py
@@ -1583,26 +1583,31 @@ class OnedriveManager:
 							foldersUnSelected.append(item["path"])
 
 		if not syncAll:
-			self.createFilterFile(foldersSelected,foldersUnSelected)
-			if not keepFolders:
-				self._manageEmptyToken()
-				if os.path.exists(self.spaceLocalFolder):
-					shutil.rmtree(self.spaceLocalFolder)
-			self.readFilterFile()
+			ret=self.createFilterFile(foldersSelected,foldersUnSelected)
+			if ret:
+				if not keepFolders:
+					self._manageEmptyToken()
+					if os.path.exists(self.spaceLocalFolder):
+						shutil.rmtree(self.spaceLocalFolder)
+				self.readFilterFile()
 		else:
+			ret=True
 			if os.path.exists(self.filterFile):
 				os.remove(self.filterFile)
 			if os.path.exists(self.filterFileHash):
 				os.remove(self.filterFileHash)
 
-		self.currentSyncConfig[0]=syncAll
-		self.currentSyncConfig[1]=foldersSelected
-		self.currentSyncConfig[2]=foldersUnSelected
-		self.folderStructBack=copy.deepcopy(self.folderStruct)
+		if ret:
+			self.currentSyncConfig[0]=syncAll
+			self.currentSyncConfig[1]=foldersSelected
+			self.currentSyncConfig[2]=foldersUnSelected
+			self.folderStructBack=copy.deepcopy(self.folderStruct)
 
-		ret=self._syncResync()
-		self._addDirectoryFile(self.spaceBasicInfo[2])	
-		return ret
+			ret=self._syncResync()
+			self._addDirectoryFile(self.spaceBasicInfo[2])	
+			return ret
+		else:
+			return ret
 
 	#def applySyncChanges
 
@@ -1628,7 +1633,7 @@ class OnedriveManager:
 		for i in range(len(foldersUnSelected)-1,-1,-1):
 			for element in foldersSelected:
 				try:
-					tmpFolder=foldersUnSelected[i]
+					tmpFolder=foldersUnSelected[i]+"/"
 					if tmpFolder in element:
 						foldersUnSelected.pop(i)
 				except:
@@ -1664,10 +1669,14 @@ class OnedriveManager:
 				
 				if os.path.exists(self.filterFile+".tmp"):
 					os.remove(self.filterFile+".tmp")
+
+			return True
 				
 		except:
 			if os.path.exists(self.filterFile+".tmp"):
 				os.rename(self.filterFile+".tmp",self.filterFile)
+
+			return False
 
 	#def createFilterFile
